test suite and make Animal pass it
I haven't developed the solution using TDD, obviously: it was just too
simple to bother. But now that I'm adding extra's I don't want to break
the original solution, hence the test suite.

One problem with the original solution was that it expected floating
point distances and broke on integers. Due to the way command line
argument was parsed, the method never received an integer, hence
the problem haven't manifested until now.